Style: spaces.  Replaced size with texture size. Removed duplicate Math:: call. Renamed objectNodeContent with node. Added update virtual method: update renderer. Replaced ISize parameter int with double.

+++ MogeLib/inc/ITexture.h
@@ -12,8 +12,8 @@ namespace Moge
 		ITexture();
 		virtual ~ITexture();
 		virtual const char* getPath()const = 0;
-		virtual Math::ISize<int>& getSize() = 0;
-		virtual const Math::ISize<int>& getSize()const = 0;
+		virtual Math::ISize<double>& getSize() = 0;
+		virtual const Math::ISize<double>& getSize()const = 0;
 	protected:
 	private:
 	};
